+++ Tests/Entity/PageTest.php
@@ -22,7 +22,6 @@ class PageTest extends AbstractTestCase
         $page = new Page();
         return $page
             ->setHomepage(true)
-            ->setEnabled(false)
             ->setSlug('home')
             ->setTitle('My homepage')
             ->setHost('localhost')
@@ -46,7 +45,7 @@ class PageTest extends AbstractTestCase
 
         $this->assertEquals($homepage->getTitle(), (string) $homepage);
 
-        $this->assertFalse($homepage->isEnabled());
+        $this->assertFalse($homepage->isEnabled()); // Base value in entity
         $this->assertFalse($homepage->isDeleted());
         $this->assertTrue($homepage->isHomepage());
         $this->assertEquals('localhost', $homepage->getHost());
@@ -79,8 +78,10 @@ class PageTest extends AbstractTestCase
 
         $this->assertNotNull($homepage);
 
-        $em->remove($homepage);
-        $em->flush();
+        if (null !== $homepage) {
+            $em->remove($homepage);
+            $em->flush();
+        }
 
         $homepage = $em->getRepository(get_class($homepage))->findOneBy(array('id' => $homepage->getId()));
